I have change the component name from the LandingPage to Title and I also removed the old cold from my code.
